Remove MeshAnimation class & file, move uv anim export to material
We don't want to oversplit it, and UV animation is driven by the material / shader, so it does not care about the UV geometry (unlike morphs etc)
