+++ io_scene_nif/modules/animation/animation_export.py
@@ -43,7 +43,6 @@ from pyffi.formats.nif import NifFormat
 
 from io_scene_nif.modules.animation.material_export import MaterialAnimation
 from io_scene_nif.modules.animation.transform_export import TransformAnimation
-from io_scene_nif.modules.animation.mesh_export import MeshAnimation
 from io_scene_nif.modules.animation.object_export import ObjectAnimation
 from io_scene_nif.modules.animation.texture_export import TextureAnimation
 from io_scene_nif.modules.object.block_registry import block_store
@@ -92,7 +91,6 @@ class Animation:
         self.mat_anim = MaterialAnimation()
         self.txt_anim = TextureAnimation(parent)
         self.transform = TransformAnimation(parent)
-        self.mesh_anim = MeshAnimation()
         self.fps = bpy.context.scene.render.fps
